Divide code and more abstract types
* New types:
Ideally we can use an index outside of an array in a way that doesn't
require strict subtyping of AbstractUnitRange. That makes things a bit
trickier though because a lot of function in Base depend on this. To
make this simpler this is going to involve a specific subtype of
AbstractIndex that will be forced through these methods and guarantees
that the keys are AbstractUnitRange

* Flexible API for combining indices:
There are a lot of places that AxisArrays got stuck because it's
difficult to find a one fits all behavior for combining indices in
different ways. Therefore, `cat`, `merge`, `join`, etc are going to be
split into three distinct steps where subtypes of AbstractIndex can
inherit broadly agreed upon behavior or customize.